pyproject.toml
[tool.poetry]
name = "apexweb"
version = "0.1.0"
description = "Apex Cabinetry's website"
authors = ["Tyler Estes <estysdesu@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1.2"
python-dotenv = "^0.13.0"
flask-bootstrap4 = "^4.0.2"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
mypy = "^0.770"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
